# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AUG22pt-ep1-dissolution2",
    "description": "<b>Swagger Doc EP Noroff Backend API</b>."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "scheme": "bearer",
      "description": "Bearer token"
    }
  },
  "paths": {
...
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Admin get all Users with there History. User logged inn, se there DataHistory",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, data: { users: \"users found\", users, DataHistory: DataHistory }"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "description": "This is not Implemented her, User are registrated with /Auth/Registrate",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {
          "403": {
            "description": "statusCode\": 403, result: {error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode\": 500, result: {error: \"POST for Users is not implemented\"}"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Admin can change/update User Privileges to Admin: \"Rights\" or Back to User: \"Rights\"",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UsersPut"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, result: { message: \"User privileges updated\" }"
          },
          "403": {
            "description": "statusCode: 403, result: { error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode\": 500, result: { error: \"Privileges On Admin cant be changed!!\"}"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Admin can Delete user. Senario: User is not active, havent bouth anything - no history",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UsersDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, result: { message: \"User is deleted from the database..\" }"
          },
          "403": {
            "description": "statusCode: 403, result: { error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode\": 500, result: { error: \"User is active, restricted to delete..\"}"
          }
        }
      }
    },
    "/roles/": {
      "get": {
        "tags": [
          "Role"
        ],
        "description": "Admin get all roles in the database",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "statusCode: 200, data: { roles: \"Roles found\", roles }"
          },
          "403": {
            "description": "statusCode: 403, result: {error: \"Privileges Admin rights [Required] !!\"}"
          }
        }
      },
      "post": {
        "tags": [
          "Role"
        ],
        "description": "All tables that are created must have CRUD endpoinst - Not implemented",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "403": {
            "description": "statusCode\": 403, result: {error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode\": 500, result: {error: \"POST for Roles is not implemented\"}"
          }
        },
        "deprecated": true
      },
      "put": {
        "tags": [
          "Role"
        ],
        "description": "All tables that are created must have CRUD endpoinst - Not implemented",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "403": {
            "description": "statusCode\": 403, result: {error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode\": 500, result: {error: \"PUT for Roles is not implemented\"}"
          }
        },
        "deprecated": true
      },
      "delete": {
        "tags": [
          "Role"
        ],
        "description": "All tables that are created must have CRUD endpoinst - Not implemented",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "403": {
            "description": "statusCode\": 403, result: {error: \"Privileges Admin rights [Required] !!\"}"
          },
          "500": {
            "description": "statusCode\": 500, result: {error: \"DELETE for Roles is not implemented\"}"
          }
        },
        "deprecated": true
      }
    },
...
  },
  "definitions": {
...
    "UsersPut": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "number",
          "example": 2
        },
        "roleName": {
          "type": "string",
          "example": "Admin"
        }
      },
      "required": [
        "userId",
        "roleName"
      ]
    },
    "UsersDelete": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "userId"
      ]
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}